fix: Adds type definition to event to avoid compiler errors (#114)
This was a very minor thing I noticed, didn't think opening an issue would be needed for something so small. I'm happy to go do so if that's still required.

I was going through the "getting started" tutorial while doing everything inside an Angular project with strict typing set to false. I noticed that this specific event was defaulting to the GameEvent type in my compiler, resulting in a compiler error when trying to access the contact property.  

## Changes:

I could see this easily tripping up someone who might not know what's going on so I've just added explicit typing.